implemented role verification for user specific operations
For example, only user with id 1 can add a recipe to the user with id 1. Same for update, delete and viewing favorite recipes.

+++ src/main/java/com/mateipruteanu/recipeapp/controllers/UserController.java
@@ -1,5 +1,6 @@
 package com.mateipruteanu.recipeapp.controllers;
 
+import com.mateipruteanu.recipeapp.dto.UserDTO;
 import com.mateipruteanu.recipeapp.models.Ingredient;
 import com.mateipruteanu.recipeapp.models.Recipe;
 import com.mateipruteanu.recipeapp.models.RecipeIngredient;
@@ -7,8 +8,11 @@ import com.mateipruteanu.recipeapp.models.User;
 import com.mateipruteanu.recipeapp.repositories.IngredientRepository;
 import com.mateipruteanu.recipeapp.repositories.RecipeRepository;
 import com.mateipruteanu.recipeapp.repositories.UserRepository;
+import com.mateipruteanu.recipeapp.token.TokenRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,8 +27,16 @@ public class UserController {
     private RecipeRepository recipeRepository;
     @Autowired
     private IngredientRepository ingredientRepository;
+    @Autowired
+    private TokenRepository tokenRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    public UserController(PasswordEncoder passwordEncoder) {
+        this.passwordEncoder = passwordEncoder;
+    }
 
     @GetMapping("/")
+    @Secured("ROLE_ADMIN")
     public List<User> getUsers() {
         return userRepository.findAll();
     }
@@ -32,10 +44,13 @@ public class UserController {
 
     // CRUD operations for users
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUser(@PathVariable long id) {
-        if(userRepository.findById(id).isPresent()) {
-            return ResponseEntity.ok(userRepository.findById(id).get());
+//    should only be accessible by admin or the user with the id
+    public ResponseEntity<UserDTO> getUser(@RequestHeader(value="Authorization") String token, @PathVariable long id) {
+        if(userRepository.findById(id).isPresent() && isTokenIDSameAsID(token, id)) {
+            UserDTO userDTO = new UserDTO(userRepository.findById(id).get());
+            return ResponseEntity.ok(userDTO);
         }
+
         return ResponseEntity.notFound().build();
     }
 
@@ -49,11 +64,11 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<String> updateUser(@PathVariable long id, @RequestBody User user) {
-        if(userRepository.findById(id).isPresent()) {
+    public ResponseEntity<String> updateUser(@RequestHeader(value="Authorization") String token, @PathVariable long id, @RequestBody User user) {
+        if(userRepository.findById(id).isPresent() && isTokenIDSameAsID(token, id)) {
             User existingUser = userRepository.findById(id).get();
             existingUser.setUsername(user.getUsername());
-            existingUser.setPassword(user.getPassword());
+            existingUser.setPassword(passwordEncoder.encode(user.getPassword()));
             userRepository.save(existingUser);
             return ResponseEntity.ok("Updated");
         }
@@ -61,13 +76,29 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable long id) {
-        if(userRepository.findById(id).isPresent()) {
+    public ResponseEntity<String> deleteUser(@RequestHeader(value="Authorization") String token, @PathVariable long id) {
+        if(userRepository.findById(id).isPresent() && isTokenIDSameAsID(token, id)) {
             userRepository.deleteById(id);
             return ResponseEntity.ok("Deleted user with id " + id);
         }
         return ResponseEntity.notFound().build();
     }
+
+
+    //    checks if user from token has the same id as the user from the argument
+    public boolean isTokenIDSameAsID(String token, long id) {
+        token = token.substring(7);
+
+        if(tokenRepository.findByToken(token).isPresent()) {
+            long userIdFromToken = tokenRepository.findByToken(token).get().getUser().getId();
+            if(userRepository.findById(id).isPresent()) {
+                return id == userIdFromToken;
+            }
+        }
+        return false;
+    }
+
+
 
 
     //User added recipes list
@@ -102,10 +133,9 @@ public class UserController {
 
 
     @PostMapping("/{userId}/recipes")
-    public ResponseEntity<String> addRecipeToUser(@PathVariable long userId, @RequestBody Recipe recipe) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
+    public ResponseEntity<String> addRecipeToUser(@RequestHeader(value="Authorization") String token, @PathVariable long userId, @RequestBody Recipe recipe) {
+        if(userRepository.findById(userId).isPresent() && isTokenIDSameAsID(token, userId)) {
+            User user = userRepository.findById(userId).get();
             for (RecipeIngredient recipeIngredient : recipe.getRecipeIngredients()) {
                 Ingredient ingredient = recipeIngredient.getIngredient();
                 if(ingredientRepository.findByName(ingredient.getName()) != null) {
@@ -123,17 +153,18 @@ public class UserController {
             userRepository.save(user);
             return ResponseEntity.ok("Recipe added to users list");
         }
+
         return ResponseEntity.notFound().build();
     }
 
 
     @DeleteMapping("/{userId}/recipes/{recipeId}")
-    public ResponseEntity<String> removeRecipeFromUser(@PathVariable long userId, @PathVariable long recipeId) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
+    public ResponseEntity<String> removeRecipeFromUser(@RequestHeader(value="Authorization") String token, @PathVariable long userId, @PathVariable long recipeId) {
+        if(userRepository.findById(userId).isPresent() && isTokenIDSameAsID(token, userId)) {
+            User user = userRepository.findById(userId).get();
             List<Recipe> addedRecipes = user.getAddedRecipes();
             addedRecipes.removeIf(recipe -> recipe.getId() == recipeId);
+            recipeRepository.deleteById(recipeId);
             userRepository.save(user);
             return ResponseEntity.ok("Recipe removed from users list");
         }
@@ -143,20 +174,18 @@ public class UserController {
 
     //Favorites List
     @GetMapping("/{userId}/favorites")
-    public ResponseEntity<List<Recipe>> getUserFavoriteRecipes(@PathVariable long userId) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
+    public ResponseEntity<List<Recipe>> getUserFavoriteRecipes(@RequestHeader(value="Authorization") String token, @PathVariable long userId) {
+        if(userRepository.findById(userId).isPresent() && isTokenIDSameAsID(token, userId)) {
+            User user = userRepository.findById(userId).get();
             return ResponseEntity.ok(user.getFavoriteRecipes());
         }
         return ResponseEntity.notFound().build();
     }
 
     @GetMapping("/{userId}/favorites/{recipeId}")
-    public ResponseEntity<Recipe> getUserFavoriteRecipe(@PathVariable long userId, @PathVariable long recipeId) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
+    public ResponseEntity<Recipe> getUserFavoriteRecipe(@RequestHeader(value="Authorization") String token, @PathVariable long userId, @PathVariable long recipeId) {
+        if(userRepository.findById(userId).isPresent() && isTokenIDSameAsID(token, userId)) {
+            User user = userRepository.findById(userId).get();
             List<Recipe> favoriteRecipes = user.getFavoriteRecipes();
 
             Optional<Recipe> optionalRecipe;
@@ -172,10 +201,9 @@ public class UserController {
     }
 
     @PostMapping("/{userId}/favorites")
-    public ResponseEntity<String> addRecipeToFavorites(@PathVariable long userId, @RequestBody Recipe recipe) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
+    public ResponseEntity<String> addRecipeToFavorites(@RequestHeader(value="Authorization") String token, @PathVariable long userId, @RequestBody Recipe recipe) {
+        if(userRepository.findById(userId).isPresent() && isTokenIDSameAsID(token, userId)) {
+            User user = userRepository.findById(userId).get();
             Optional<Recipe> optionalRecipe = recipeRepository.findById(recipe.getId());
             if(optionalRecipe.isPresent()) {
                 Recipe recipe1 = optionalRecipe.get();
@@ -188,16 +216,17 @@ public class UserController {
     }
 
     @DeleteMapping("/{userId}/favorites/{recipeId}")
-    public ResponseEntity<String> removeRecipeFromFavorites(@PathVariable long userId, @PathVariable long recipeId) {
-        Optional<User> optionalUser = userRepository.findById(userId);
-        if (optionalUser.isPresent()) {
-            User user = optionalUser.get();
+    public ResponseEntity<String> removeRecipeFromFavorites(@RequestHeader(value="Authorization") String token, @PathVariable long userId, @PathVariable long recipeId) {
+        if(userRepository.findById(userId).isPresent() && isTokenIDSameAsID(token, userId)) {
+            User user = userRepository.findById(userId).get();
             List<Recipe> favoriteRecipes = user.getFavoriteRecipes();
             favoriteRecipes.removeIf(recipe -> recipe.getId() == recipeId);
             userRepository.save(user);
             return ResponseEntity.ok("Recipe removed from favorites list for user" + user.getUsername());
         }
+
         return ResponseEntity.notFound().build();
     }
+
 
 }
